+++ lib/src/country_button_style.dart
@@ -15,9 +15,6 @@ class CountryButtonStyle {
   /// Whether to show the country flag in the country button
   final bool showFlag;
 
-  /// Whether to apply greyscale filter on flag icon if control is disabled
-  final bool grayScaleFlagOnDisabled;
-
   /// Whether to show Dial Code in the country button
   /// setting this to false will hide, for example, (+1)
   final bool showDialCode;
@@ -33,7 +30,6 @@ class CountryButtonStyle {
     this.padding,
     this.flagSize = 20,
     this.showFlag = true,
-    this.grayScaleFlagOnDisabled = true,
     this.showDialCode = true,
     this.showIsoCode = false,
     this.showDropdownIcon = true,
@@ -43,7 +39,6 @@ class CountryButtonStyle {
     TextStyle? textStyle,
     EdgeInsets? padding,
     double? flagSize,
-    bool? grayScaleFlagOnDisabled,
     bool? showFlag,
     bool? showDialCode,
     bool? showIsoCode,
@@ -53,8 +48,6 @@ class CountryButtonStyle {
       textStyle: textStyle ?? this.textStyle,
       padding: padding ?? this.padding,
       flagSize: flagSize ?? this.flagSize,
-      grayScaleFlagOnDisabled:
-          grayScaleFlagOnDisabled ?? this.grayScaleFlagOnDisabled,
       showFlag: showFlag ?? this.showFlag,
       showDialCode: showDialCode ?? this.showDialCode,
       showIsoCode: showIsoCode ?? this.showIsoCode,